update paging adapter.

+++ app/src/main/java/com/example/retrofitwithpaging/response/MovieDetailsResponse.kt
@@ -1,5 +1,6 @@
 package com.example.retrofitwithpaging.response
 
+import android.util.Log
 import com.google.gson.annotations.SerializedName
 
 data class MovieDetailsResponse(
@@ -36,7 +37,7 @@ data class MovieDetailsResponse(
     @SerializedName("release_date")
     val releaseDate: String,
     @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
     @SerializedName("runtime")
     val runtime: Int,
     @SerializedName("spoken_languages")
